% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{analyselist}
\alias{analyselist}
\title{analyselist}
\usage{
analyselist(id, datlist, varlist)
}
\arguments{
\item{id}{patient identifier}

\item{datlist}{imputed dataset of M imputations}

\item{varlist}{list of derived variables ,varlist <- c("fev.2","fev.4","fev.8","fev.12","base")}
}
\value{
printout of descriptve stats
}
\description{
find descriptive stats on the  M imputed data set
}
\details{
select on patient id and find their means etc
}
\examples{
\dontrun{
varlist <- c("fev.2","fev.4","fev.8","fev.12","base")
analyselist(5099,impdataset,varlist)
}
}
