% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Causal_loop}
\alias{Causal_loop}
\title{Causal_loop}
\usage{
Causal_loop(c_mata_miss, mata_Means, MeansC, Kd)
}
\arguments{
\item{c_mata_miss}{vector of missing}

\item{mata_Means}{fill-in here}

\item{MeansC}{fill-in}

\item{Kd}{parameter that makes it between J2R and CIR, value 1 equiv to CIR}
}
\value{
mata_means
}
\description{
process Causal method
}
\details{
This is based on Suzie Cro's Stata program
}
