% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CIR_loop}
\alias{CIR_loop}
\title{CIR_loop}
\usage{
CIR_loop(c_mata_miss, mata_Means, MeansC)
}
\arguments{
\item{c_mata_miss}{vector of missing}

\item{mata_Means}{fill-in here}

\item{MeansC}{fill-in}
}
\value{
mata_means
}
\description{
process CIR method
}
\details{
This is based on Suzie Cro's Stata program
}
