% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{getimpdatasets}
\alias{getimpdatasets}
\title{getimpdatasets}
\usage{
getimpdatasets(varlist)
}
\arguments{
\item{varlist}{list of data containing imputed values from the M pattern groups}
}
\value{
impdataset
}
\description{
to obtain the M imputed data set from the output list into one dataset
}
\details{
This combines the imputations found from the M pattern groups
}
