% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess.R
\name{ifmethodindiv}
\alias{ifmethodindiv}
\title{ifmethodindiv}
\usage{
ifmethodindiv(
  methodindiv,
  mg,
  m,
  M,
  paramBiglist,
  i,
  treatvar,
  c_mata_nonmiss,
  c_mata_miss,
  mata_miss,
  mata_nonmiss
)
}
\arguments{
\item{methodindiv}{c(method,reference) cols}

\item{mg}{pattern lookup table}

\item{m}{where we are in the imputations}

\item{M}{number of total imputations.}

\item{paramBiglist}{list of Beta and Sigma parameters from mcmc}

\item{i}{in loop throug mg rows}

\item{treatvar}{treatment group}

\item{c_mata_nonmiss}{0,1 vector of nonmissing}

\item{c_mata_miss}{0,1 vector of missing}

\item{mata_miss}{position of missing values in repeated time visits}

\item{mata_nonmiss}{positions of nonmissing values}
}
\value{
mata_means
}
\description{
alternative logic for individual method
}
\details{
checks methodindiv not null
}
