% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impdataCausal}
\alias{impdataCausal}
\title{impdataCausal.}
\format{
An object of class \code{data.frame} with 1098 rows and 9 columns.
}
\usage{
impdataCausal
}
\description{
A dataset containing imputed values obtained from mimix on asthma data

@format A data frame with 1098 rows and 9 columns
\describe{
\item{.imp}{imputation number}
\item{base}{covariate}
\item{fev.2}{response at time 2}
\item{fev.4}{response at time 4}
\item{fev.8}{response at time 8}
\item{fev.12}{response at time 12}
\item{patt}{missing data pattern}
\item{treat}{treatment group}
\item{.id}{case id}
}
}
\keyword{datasets}
