% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LMCF_loop}
\alias{LMCF_loop}
\title{LMCF_loop}
\usage{
LMCF_loop(c_mata_miss, mata_Means)
}
\arguments{
\item{c_mata_miss}{vector of missing}

\item{mata_Means}{fill-in here}
}
\value{
mata_means
}
\description{
process LMCF method
}
\details{
This is based on Suzie Cro's Stata program
}
