% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{mi_impute}
\alias{mi_impute}
\title{mi_impute}
\usage{
mi_impute(data, idvar, timevar, depvar, covar)
}
\arguments{
\item{data}{data set}

\item{idvar}{patient id}

\item{timevar}{time point for repeated measure}

\item{depvar}{dependent variable}

\item{covar}{covariates and base depvar must be complete}
}
\value{
list of variables
}
\description{
create a vector defining imputation variables
}
\details{
similar to mi_impute in mimix Stata programx
}
