% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{mimix}
\alias{mimix}
\title{mimix}
\usage{
mimix(
  data,
  covar = NULL,
  depvar,
  treatvar,
  idvar,
  timevar,
  M = 1,
  refer = NULL,
  meth = NULL,
  seedval = 101,
  priorvar = "jeffreys",
  burnin = 1000,
  bbetween = NULL,
  methodindiv = NULL,
  delta = NULL,
  Kd = NULL
)
}
\arguments{
\item{data}{datset in wide format}

\item{covar}{covariates and base depvar must be complete (no missing vaules)}

\item{depvar}{dependent variable}

\item{treatvar}{treatment group , recoded to 1,2,..}

\item{idvar}{patient id}

\item{timevar}{time point for repeated measure}

\item{M}{number of imputations}

\item{refer}{reference group for j2r,cir,cr mrthods}

\item{meth}{RBI method}

\item{seedval}{seed value to obtain same outputs}

\item{priorvar}{prior,defualt Jeffries, also uniform or ridge}

\item{burnin}{burnin value}

\item{bbetween}{value between iterations in mcmc}

\item{methodindiv}{2 element vector designating variables in data specifying individual method and reference group}

\item{delta}{vector of delta values to add onto imputed values (non-mandatory)}

\item{Kd}{Causal constant for use with Causal method}
}
\value{
impdataset the m impute data-sets appended to the "missing values" data-set in wide format
}
\description{
main wrapper for running mimix
}
\details{
This is based on Suzie Cro's Stata program
}
