% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi_package.R
\docType{package}
\name{mimix_package}
\alias{mimix_package}
\title{mimix: A package porting the Stata mimix command}
\description{
The mimix package provides the functionality of the Stata package plus
delta and causal methods
}
\section{Comparison with Stata}{

The mimix functions add the causal method and delta extra to the Stata version  ...
}

\section{Comparison with SAS}{

The mimix functions ...
The mimix functions  causal method and delta extra to the Stata version  ...
are similar to the SAS functions found in ..
}

