% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{pass2Loop}
\alias{pass2Loop}
\title{Performs the imputation for the specified method after MAR ran}
\usage{
pass2Loop(
  Imp_Interims,
  method,
  mg,
  ntreat,
  depvar,
  covar,
  treatvar,
  reference,
  trtgp,
  mata_Obs,
  mata_all_newlist,
  paramBiglist,
  idvar,
  flag_indiv,
  M,
  delta,
  dlag,
  K0,
  K1
)
}
\arguments{
\item{Imp_Interims}{Interim cases}

\item{method}{\itemize{
\item Specified model to run Reference-based imputation method
}}

\item{mg}{the summary table based on missing data patern}

\item{ntreat}{vector of treatment groups}

\item{depvar}{response variable}

\item{covar}{covariate variable(s)}

\item{treatvar}{Treatment group, coded 1,2,..}

\item{reference}{Reference group for J2R, CIR, CR methods}

\item{trtgp}{treatmet grp}

\item{mata_Obs}{raw data with interims imputed}

\item{mata_all_newlist}{raw data with interims imputed in list}

\item{paramBiglist}{list of MCMC beta and Sigma parameters}

\item{idvar}{Participant id}

\item{flag_indiv}{flag whether specified individual column in data}

\item{M}{number of imputations}

\item{delta}{vector of delta values to add onto imputed values (non-mandatory) (a's in Rogers paper),length as number of time points}

\item{dlag}{vector of delta values to add onto imputed values (non-mandatory) (b's in Rogers paper),length as number of time points}

\item{K0}{Causal constant for use with Causal method}

\item{K1}{exponential decaying Causal constant for use with Causal method}
}
\value{
impdataset the M imputed data-sets appended to the "missing values" data-set in wide format
}
\description{
2nd pass for specified method after 1st pass MAR ran
}
\details{
reads the summary table based on missing data pattern- mg  mimix_group

reflects the pattern and treatment group configuration of the raw data

then acts as a looping mechanism, norm2 is used as MCMC multivariate normal
}
\examples{
\dontrun{
testpass2impdatset<- pass2Loop(Imp_Interims,method,mg,ntreat,depvar,treatvar,reference,trtgp,mata_Obs,mata_all_newlist,paramBiglist,idvar,flag_indiv,M,delta,K0,K1)
}
}
