% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess.R
\name{preprodata}
\alias{preprodata}
\title{pre-process long longitudinal data into wide format}
\usage{
preprodata(
  data,
  covar,
  depvar,
  treatvar,
  idvar,
  timevar,
  M,
  reference,
  method = NULL
)
}
\arguments{
\item{data}{data in long format}

\item{covar}{covariates and base depvar complete}

\item{depvar}{dependent variable}

\item{treatvar}{treatment group}

\item{idvar}{patient id}

\item{timevar}{time variable for repeated visit}

\item{M}{number imputations}

\item{reference}{reference group}

\item{method}{RBI method}
}
\value{
list of outputs
}
\description{
process data into wide format for group method
}
\details{
checks method finds missingness pattern
}
