% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess.R
\name{preproIndivdata}
\alias{preproIndivdata}
\title{pre-process long longitudinal data into wide format for individual-specific}
\usage{
preproIndivdata(
  data,
  covar,
  depvar,
  treatvar,
  idvar,
  timevar,
  M,
  reference = NULL,
  method = NULL,
  methodvar,
  referencevar
)
}
\arguments{
\item{data}{data in long format}

\item{covar}{covariates and base depvar complete}

\item{depvar}{dependent variable}

\item{treatvar}{treatment group}

\item{idvar}{patient id}

\item{timevar}{time variable for repeated visit}

\item{M}{number imputations}

\item{reference}{reference group must be NULL}

\item{method}{RBI method  must be NULL}

\item{methodvar}{column location in data specifying individual RBI methods}

\item{referencevar}{column location in data specifying individual reference group for RBI method}
}
\value{
list of outputs
}
\description{
process data into wide format for individual-specified method
}
\details{
checks methodvar finds missingness pattern
}
