% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pttestf}
\alias{pttestf}
\title{pttestf}
\usage{
pttestf(n1, n2, mn1, sd1, mn2, sd2)
}
\arguments{
\item{n1}{sample size  group1}

\item{n2}{samplesize group}

\item{mn1}{mean value in sample 1}

\item{sd1}{standard deviation in sample 1}

\item{mn2}{mean value in sample 2}

\item{sd2}{stadard deviation sample 2}
}
\value{
p value
}
\description{
significance test between mean of 2 groups
}
\details{
This is based on ..
}
