% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefBasedMI-package.R
\docType{package}
\name{RefBasedMI-package}
\alias{RefBasedMI-package}
\title{RefBasedMI: Reference-Based Imputation for Longitudinal Clinical Trials with Protocol Deviation}
\description{
Imputation of missing numerical outcomes for a longitudinal trial with protocol deviations. The package uses distinct treatment arm-based assumptions for the unobserved data, following the general algorithm of Carpenter, Roger, and Kenward (2013) \doi{10.1080/10543406.2013.834911}, and the causal model of White, Royes and Best (2020) \doi{10.1080/10543406.2019.1684308}. Sensitivity analyses to departures from these assumptions can be done by the Delta method of Roger. The program uses the same algorithm as the 'mimix' 'Stata' package written by Suzie Cro, with additional coding for the causal model and delta method. The reference-based methods are jump to reference (J2R), copy increments in reference (CIR), copy reference (CR), and the causal model, all of which must specify the reference treatment arm. Other methods are missing at random (MAR) and the last mean carried forward (LMCF).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/choxos/RefbasedMI}
  \item Report bugs at \url{https://github.com/choxos/RefbasedMI/issues}
}

}
\author{
\strong{Maintainer}: Matteo Quartagno \email{m.quartagno@ucl.ac.uk}

Authors:
\itemize{
  \item Kevin McGrath \email{kkevinmcgrath@yahoo.co.uk}
  \item Ahmad Sofi-Mahmudi \email{a.sofimahmudi@gmail.com} (\href{https://orcid.org/0000-0001-6829-0823}{ORCID})
}

}
\keyword{internal}
