% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{regressimp}
\alias{regressimp}
\title{regressimp}
\usage{
regressimp(dataf, regmodel)
}
\arguments{
\item{dataf}{data-frame}

\item{regmodel}{regression model specfication}
}
\value{
estimates of regression coefficients
}
\description{
run regression on M imputed data set, combinging as Rubin's rules
}
\details{
This is approach followed from  norm2 user manual
}
