% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{Runmimix}
\alias{Runmimix}
\title{Runmimix}
\usage{
Runmimix(
  data,
  covar,
  depvar,
  treatvar,
  idvar,
  timevar,
  M = 1,
  refer,
  meth,
  seedval = 101,
  priorvar,
  burnin = 1000,
  bbetween = NULL,
  methodindiv
)
}
\arguments{
\item{data}{datset in wide format}

\item{covar}{covariates and base depvar must be complete (no missing vaules)}

\item{depvar}{dependent variable}

\item{treatvar}{treatment group , recoded to 1,2,..}

\item{idvar}{patient id}

\item{timevar}{time point for repeated measure}

\item{M}{number of imputations}

\item{refer}{reference group for j2r,cir,cr mrthods}

\item{meth}{RBI method}

\item{seedval}{seed value to obtain same outputs}

\item{priorvar}{prior,defualt Jeffries, also uniform or ridge}

\item{burnin}{burnin value}

\item{bbetween}{value between iterations in mcmc}

\item{methodindiv}{columns in dat for individual method and reference group}
}
\value{
impdataset
}
\description{
main wrapper for running mimix
}
\details{
This is based on Suzie Cro's Stata program
}
