% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acupuncture}
\alias{acupuncture}
\title{acupuncture.}
\format{
An object of class \code{data.frame} with 802 rows and 11 columns.
}
\usage{
acupuncture
}
\description{
A dataset containing trial results of a randomised, double-blind, parallel-group studies comparing active treatment with placebo
}
\details{
@format A data frame with 688 rows and 14 columns
 \describe{
 \item{id}{}
 \item{time}{}
 \item{age}{}
 \item{sex}{}
 \item{migraine}{}
 \item{chronicity}{}
 \item{practice_id}{}
 \item{treat}{}
 \item{head_base}{}
 \item{head}{}
 \item{withdrawal_reason}{}
 }
}
\keyword{datasets}
