% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acupuncture}
\alias{acupuncture}
\title{Sample data: acupuncture trial}
\format{
A data frame with 802 rows and 11 columns
\describe{
\item{id}{}
\item{time}{}
\item{age}{}
\item{sex}{}
\item{migraine}{}
\item{chronicity}{}
\item{practice_id}{}
\item{treat}{}
\item{head_base}{covariate}
\item{head}{outcome variable}
\item{withdrawal_reason}{}
}
}
\usage{
acupuncture
}
\description{
A data set containing results of a randomised, double-blind, parallel-group
comparing active treatment with placebo
The primary outcome is head, measured at time 3 and 12
}
\keyword{datasets}
