% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acupuncture}
\alias{acupuncture}
\title{Sample data: acupuncture trial}
\format{
A data frame with 802 rows and 11 columns
\describe{
\item{id}{}
\item{time}{}
\item{age}{}
\item{sex}{}
\item{migraine}{}
\item{chronicity}{}
\item{practice_id}{}
\item{treat}{}
\item{head_base}{covariate}
\item{head}{dependent variable}
\item{withdrawal_reason}{}
}
}
\usage{
acupuncture
}
\description{
A data set containing results of a randomised, double-blind, parallel-group
comparing active treatment with placebo
The primary outcome is head, measured at time 3 and 12
}
\examples{
\dontrun{
 impCausalref1 <- mimix(data=acupuncture,covar=c("head_base","sex"),depvar=head,treatvar=treat,
       idvar=id,timevar=time,
       method=Causal,reference=1,K0=1,K1=0.5,M=5,seed=54321)
 library(mice)         
 fit<-with(as.mids(impCausalref1), lm(head.12~treat+head_base+sex))  
 summary(pool(fit))    
}
}
\keyword{datasets}
