% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AddDelta}
\alias{AddDelta}
\title{AddDelta}
\usage{
AddDelta(vec_tst, ncovar, mata_imp, delta, dlag)
}
\arguments{
\item{vec_tst}{vector of visit names}

\item{ncovar}{number covariates}

\item{mata_imp}{the imputed values (as well as the complete)}

\item{delta}{vector of delta values (a values in Roger's) for each vist time}

\item{dlag}{vector of dlag values (b values in Roger's) for each vist time}
}
\value{
mata_imp the adjusted imputed vaues (and unadjusted non-missing)
}
\description{
add delta's to imputed values
}
\details{
adding delta values after wthdrawal
}
