% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antidepressant}
\alias{antidepressant}
\title{Sample data: antidepressant trial}
\format{
dataframe containing 688 rows and 14 columns
\describe{
\item{PATIENT.NUMBER}{}
\item{HAMA.TOTAL}{}
\item{PGI_IMPROVEMENT}{}
\item{VISIT...VISIT.3.DATE}{}
\item{VISIT.NUMBER}{}
\item{TREATMENT.NAME}{}
\item{PATIENT.SEX}{}
\item{POOLED.INVESTIGATOR}{}
\item{basval}{}
\item{HAMD17.TOTAL}{outcome variable}
\item{change}{}
\item{miss_flag}{}
\item{methodcol}{individual-specific method}
\item{referencecol}{individual-specific reference arm}
}
}
\usage{
antidepressant
}
\description{
A data set containing antidepressant trial data as described in paper
by White,Royes,Best (2019)
The primary outcome is HAMD17.TOTAL measured at visit number 4,5,6,7.
}
\keyword{datasets}
