% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antidepressant}
\alias{antidepressant}
\title{Sample data: antidepressant trial}
\format{
dataframe containing 688 rows and 14 columns
\describe{
\item{PATIENT.NUMBER}{}
\item{HAMA.TOTAL}{}
\item{PGI_IMPROVEMENT}{}
\item{VISIT...VISIT.3.DATE}{}
\item{VISIT.NUMBER}{}
\item{TREATMENT.NAME}{}
\item{PATIENT.SEX}{}
\item{POOLED.INVESTIGATOR}{}
\item{basval}{}
\item{HAMD17.TOTAL}{outcome variable}
\item{change}{}
\item{miss_flag}{}
\item{methodcol}{individual-specific method}
\item{referencecol}{individual-specific reference arm}
}
}
\usage{
antidepressant
}
\description{
A data set containing antidepressant trial data as described in paper
by White,Royes,Best (2019)
The primary outcome is HAMD17.TOTAL measured at visit number 4,5,6,7.
}
\examples{
\donttest{
 # Run with  covariates "basval" and "PATIENT.SEX" using columns within data to specify
 # method and reference indivdually specified columns 
 impIndiv <- RefBasedMI(data=antidepressant,covar=c(basval,PATIENT.SEX),depvar=HAMD17.TOTAL,
     treatvar=TREATMENT.NAME,idvar=PATIENT.NUMBER,
     timevar=VISIT.NUMBER,methodvar="methodcol",referencevar="referencecol",M=5,seed=54321)        
 library(mice)
 fit<-with(data= as.mids(subset(impIndiv,VISIT.NUMBER==7)),
               lm(HAMD17.TOTAL~TREATMENT.NAME+basval+PATIENT.SEX))
 summary(pool(fit))  
 impantdep <- RefBasedMI(data=antidepressant,covar=c(basval,PATIENT.SEX),depvar=HAMD17.TOTAL,
         treatvar=TREATMENT.NAME,idvar=PATIENT.NUMBER,
         timevar=VISIT.NUMBER,method="J2R",reference=1,M=5,seed=54321)
 fitdep21<-with(data= as.mids(subset(impantdep,VISIT.NUMBER==7)),
              lm(HAMD17.TOTAL~TREATMENT.NAME+basval))
 summary(pool(fitdep21)) 
}
}
\keyword{datasets}
