% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asthma}
\alias{asthma}
\title{Sample data: asthma trial}
\format{
A data frame containing 732 rows and 5 columns
\describe{
\item{id}{patient identifier}
\item{time}{}
\item{treat}{}
\item{base}{covariate}
\item{fev}{outcome variable}
}
}
\usage{
asthma
}
\description{
A data set containing  asthma trial data as used in the Stata mimix help file
The primary outcome variable is fev, measured at 2,4,8,12 weeks
}
\examples{
\dontrun{
 impJ2Rridge<-(mimix(data=asthma,covar=c("base"),depvar=fev,treatvar=treat,idvar=id,timevar=time,
    method="J2R",reference=1,delta=c(0.5,0.5,1,1 ),M=5,seed=101,prior=ridge,burnin=1000)
 library(mice) 
 fitJ2R<-with(data= as.mids(subset(impJ2Rridge[[2]],time==12)),lm(fev~treat+base))    
 summary(pool(fitJ2R))    
}     
}
\keyword{datasets}
