% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asthma}
\alias{asthma}
\title{asthma.}
\format{
An object of class \code{data.frame} with 732 rows and 5 columns.
}
\usage{
asthma
}
\description{
A dataset containg  asthma trial data as used in the Stata mimix help file
}
\details{
@format A data frame containing 732 rows and 5 columns
\describe{
\item{id}{}
\item{time}{}
\item{treat}{}
\item{base}{}
\item{fev}{}
}
}
\keyword{datasets}
