% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{getimpdatasets}
\alias{getimpdatasets}
\title{to obtain the M'th imputed data set from the output list into one dataset}
\usage{
getimpdatasets(varlist)
}
\arguments{
\item{varlist}{list of data containing imputed values from the M pattern groups}
}
\value{
impdatasets
}
\description{
to append the M imputed data sets wit hte original unimputed data
}
\details{
This combines the imputations found from the M pattern groups
}
