% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess.R
\name{ifmethodindiv}
\alias{ifmethodindiv}
\title{ifmethodindiv}
\usage{
ifmethodindiv(
  methodvar,
  referencevar,
  mg,
  m,
  M,
  paramBiglist,
  i,
  treatvar,
  c_mata_nonmiss,
  c_mata_miss,
  mata_miss,
  mata_nonmiss,
  K0,
  K1
)
}
\arguments{
\item{methodvar}{individual method col}

\item{referencevar}{individual reference col}

\item{mg}{pattern lookup table}

\item{m}{where we are in the imputations}

\item{M}{number of total imputations.}

\item{paramBiglist}{list of Beta and Sigma parameters from mcmc}

\item{i}{in loop through mg rows}

\item{treatvar}{treatment group}

\item{c_mata_nonmiss}{0,1 vector of nonmissing}

\item{c_mata_miss}{0,1 vector of missing}

\item{mata_miss}{position of missing values in repeated time visits}

\item{mata_nonmiss}{positions of nonmissing values}

\item{K0}{Causal constant for use with Causal method}

\item{K1}{exponential decaying Causal constant for use with Causal method}
}
\value{
list of outputs
}
\description{
alternative logic for individual method
}
\details{
checks methodindiv not null
}
