% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess.R
\name{preproIndivdata}
\alias{preproIndivdata}
\title{preproIndivdata}
\usage{
preproIndivdata(
  data,
  covar,
  depvar,
  treatvar,
  idvar,
  timevar,
  M,
  refer = NULL,
  meth = NULL,
  methodindiv
)
}
\arguments{
\item{data}{data in long format}

\item{covar}{covariates and base depvar complete}

\item{depvar}{dependent variable}

\item{treatvar}{treatment group}

\item{idvar}{patient id}

\item{timevar}{time variable for repeated visit}

\item{M}{number imputations}

\item{refer}{reference group must be NULL}

\item{meth}{RBI method  must be NULL}

\item{methodindiv}{list of c(method, reference) column location in data specifying individual RBI methods and references}
}
\value{
list of outputs
}
\description{
process data into wide format for individual-specified method
}
\details{
checks methodindiv finds missingness pattern
}
